Fix: Compilation error about 'fm_sem'
Compilation complains about fm_sem not being defined.
This error is valid, because in our sourcedrop it is called fm_eba_sem.
Probably Quectel changed some code to the ubi driver, because mainline
defines it as fm_sem as well.
But because fm_eba_sem has been used already in many places in the code
just change these over to fm_eba_sem as well. Is the easiest fix.