Fix: Compilation error invalid operand to 'pnum'
Another code clean-up to the UBI code.
In all code pnum is just used as plain "int pnum"
For some reason in the check_mapping function it is defined as
int *pnum, which it complains about as invalid operand.
Change all instances op *pnum to plain pnum, inline with the rest of the code.